backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
alembic==1.13.1
aiosqlite==0.20.0
httpx==0.27.0
anthropic==0.28.0
python-multipart==0.0.9
pillow==10.3.0
pdf2image==1.17.0
beautifulsoup4==4.12.3
lxml==5.2.2
requests==2.32.2
pydantic==2.7.1
pydantic-settings==2.2.1
python-jose==3.3.0
aiofiles==23.2.1
